# [Numpy] FFI for linalg.eig/eigh/det/slogdet